Switching to StandardScaler
My results seem to be less accuate, and I'm wondering if that's related to the scaler that I'm using.  I tried MinMax, Standard and Robust in previous runs and was hitting 0.525 on Kaggle, but I didn't keep good records about what I did.  Trying to get back to that baseline and kicking myself for not doing more commits (although I was traveling and offline for a bunch of it... ugh.)